backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via redis
django==5.0
    # via
    #   -r base.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-timezone-field
    #   django-unfold
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
#django-celery-beat==2.1.0
    # via -r base.in
django-cors-headers==4.3.1
    # via -r base.in
django-environ==0.11.2
    # via -r base.in
django-extensions==3.2.3
    # via -r base.in
django-filter==23.5
    # via -r base.in
#django-jazzmin==2.6.0
    # via -r base.in
django-timezone-field==4.2.3
    # via django-celery-beat
django-unfold==0.20.1
    # via -r base.in
djangorestframework==3.14.0
    # via
    #   -r base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r base.in
drf-spectacular==0.27.0

gunicorn==19.9.0
mysqlclient==2.0.3
Pillow==10.0.1
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
itsdangerous
bcrypt==4.0.1
django-cors-headers
